TIP-874: move Product's events
