+++ src/Pim/Bundle/DataGridBundle/spec/Extension/MassAction/Handler/DeleteProductsMassActionHandlerSpec.php
@@ -17,7 +17,7 @@ use Pim\Bundle\DataGridBundle\Extension\MassAction\Actions\Ajax\DeleteMassAction
 use Pim\Bundle\DataGridBundle\Extension\MassAction\Event\MassActionEvent;
 use Pim\Bundle\DataGridBundle\Extension\MassAction\Event\MassActionEvents;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
-use Pim\Component\Catalog\ProductEvents;
+use Akeneo\Pim\Enrichment\Component\Product\ProductEvents;
 use Akeneo\Pim\Enrichment\Component\Product\Query\ProductQueryBuilderInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Repository\ProductMassActionRepositoryInterface;
 use Prophecy\Argument;
